Simple SHA-1 hash provider for Keycloak
