+++ .github/workflows/docker.yml
@@ -4,8 +4,8 @@ on:
       keycloak_version:
         description: Keycloak version
         required: true
-      keycloak_bcrypt_version:
-        description: Keycloak BCrypt version
+      keycloak_sha1_version:
+        description: Keycloak SHA-1 version
         required: true
         default: 1.5.1
 
@@ -19,7 +19,7 @@ jobs:
       - uses: actions/checkout@v2
       - name: Build
         run: ./gradlew jar -Pdependency.keycloak.version=${{ github.event.inputs.keycloak_version }}
-      - run: cp build/libs/keycloak-bcrypt-${{ github.event.inputs.keycloak_bcrypt_version }}.jar docker
+      - run: cp build/libs/keycloak-sha1-${{ github.event.inputs.keycloak_sha1_version }}.jar docker
       - name: Login to DockerHub
         uses: docker/login-action@v1
         with:
@@ -31,6 +31,6 @@ jobs:
           context: docker
           build-args: |
             keycloak_version=${{ github.event.inputs.keycloak_version }}
-            keycloak_bcrypt_version=${{ github.event.inputs.keycloak_bcrypt_version }}
+            keycloak_sha1_version=${{ github.event.inputs.keycloak_sha1_version }}
           push: true
-          tags: gleroy/keycloak-bcrypt:${{ github.event.inputs.keycloak_version }}_${{ github.event.inputs.keycloak_bcrypt_version }},gleroy/keycloak-bcrypt:latest
+          tags: jcschaff/keycloak-sha1:${{ github.event.inputs.keycloak_version }}_${{ github.event.inputs.keycloak_sha1_version }},jcschaff/keycloak-sha1:latest
