+++ .github/workflows/release.yml
@@ -30,6 +30,6 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         with:
           upload_url: ${{ steps.create_release.outputs.upload_url }}
-          asset_path: ./build/libs/keycloak-bcrypt-${{ github.GITHUB_REF_NAME }}.jar
-          asset_name: keycloak-bcrypt-${{ github.GITHUB_REF_NAME }}.jar
+          asset_path: ./build/libs/keycloak-sha1-${{ github.GITHUB_REF_NAME }}.jar
+          asset_name: keycloak-sha1-${{ github.GITHUB_REF_NAME }}.jar
           asset_content_type: application/java-archive
