+++ README.md
@@ -1,6 +1,8 @@
-# Keycloak BCrypt
+# Keycloak SHA-1
 
-Add a password hash provider to handle BCrypt passwords inside Keycloak.
+A minimal fork of Keycloak BCrypt to support passwords hashed with SHA-1 (rather than BCrypt)
+
+Add a password hash provider to handle SHA-1 passwords inside Keycloak.
 
 ## Build JAR
 
@@ -11,11 +13,11 @@ Add a password hash provider to handle BCrypt passwords inside Keycloak.
 ## Build Docker image
 
 ```bash
-cp build/libs/keycloak-bcrypt-${KEYCLOAK_BCRYPT_VERSION}.jar docker
+cp build/libs/keycloak-sha1-${KEYCLOAK_SHA1_VERSION}.jar docker
 docker build \
     --build-arg keycloak_version=${KEYCLOAK_VERSION} \
-    --build-arg keycloak_bcrypt_version=${KEYCLOAK_BCRYPT_VERSION} \
-    -t gleroy/keycloak-bcrypt \
+    --build-arg keycloak_sha1_version=${KEYCLOAK_SHA1_VERSION} \
+    -t jcschaff/keycloak-sha1 \
     docker
 ```
 
@@ -30,14 +32,15 @@ docker-compose up -d
 ### >= 17.0.0
 
 ```bash
-curl -L https://github.com/leroyguillaume/keycloak-bcrypt/releases/download/${KEYCLOAK_BCRYPT_VERSION}/keycloak-bcrypt-${KEYCLOAK_BCRYPT_VERSION}.jar > ${KEYCLOAK_HOME}/providers/keycloak-bcrypt-${KEYCLOAK_BCRYPT_VERSION}.jar
+curl -L https://github.com/jcschaff/keycloak-sha1/releases/download/${KEYCLOAK_SHA1_VERSION}/keycloak-sha1-${KEYCLOAK_SHA1_VERSION}.jar > ${KEYCLOAK_HOME}/providers/keycloak-sha1-${KEYCLOAK_SHA1_VERSION}.jar
 ```
 You need to restart Keycloak.
 
 ### < 17.0.0
 
 ```bash
-curl -L https://github.com/leroyguillaume/keycloak-bcrypt/releases/download/${KEYCLOAK_BCRYPT_VERSION}/keycloak-bcrypt-${KEYCLOAK_BCRYPT_VERSION}.jar > ${KEYCLOAK_HOME}/standalone/deployments/keycloak-bcrypt-${KEYCLOAK_BCRYPT_VERSION}.jar
+curl -L https://github.com/jcschaff/keycloak-sha1/releases/download/${KEYCLOAK_SHA1_VERSION}/keycloak-sha1-${KEYCLOAK_SHA1_VERSION}.jar > ${KEYCLOAK_HOME}/standalone/deployments/keycloak-sha1-${KEYCLOAK_SHA1
+_VERSION}.jar
 ```
 You need to restart Keycloak.
 
@@ -48,13 +51,13 @@ docker run \
     -e KEYCLOAK_ADMIN=${KEYCLOAK_ADMIN} \
     -e KEYCLOAK_ADMIN_PASSWORD=${KEYCLOAK_ADMIN_PASSWORD} \
     -e KC_HOSTNAME=${KC_HOSTNAME} \
-    gleroy/keycloak-bcrypt \
+    gleroy/keycloak-sha1 \
     start
 ```
 
 The image is based on [Keycloak official](https://quay.io/repository/keycloak/keycloak) one.
 
 ## How to use
-Go to `Authentication` / `Password policy` and add hashing algorithm policy with value `bcrypt`.
+Go to `Authentication` / `Password policy` and add hashing algorithm policy with value `sha1`.
 
 To test if installation works, create new user and set its credentials.
