+++ build.gradle.kts
@@ -2,8 +2,8 @@ plugins {
     java
 }
 
-group = "com.github.leroyguillaume"
-version = "1.5.1"
+group = "com.github.jcschaff"
+version = "0.9.0"
 
 repositories {
     mavenCentral()
@@ -16,13 +16,9 @@ configurations {
 }
 
 dependencies {
-    val bcryptVersion = "0.9.0"
     val jbossLoggingVersion = "3.4.1.Final"
     val keycloakVersion = project.property("dependency.keycloak.version")
     val junitVersion = "5.8.2"
-
-    // BCrypt
-    implementation("at.favre.lib:bcrypt:$bcryptVersion")
 
     // JBoss
     compileOnly("org.jboss.logging:jboss-logging:$jbossLoggingVersion")
