+++ docker/Dockerfile
@@ -1,7 +1,9 @@
-ARG keycloak_version
+ARG keycloak_version=19.0.0
 
 FROM quay.io/keycloak/keycloak:${keycloak_version}
 
-ARG keycloak_bcrypt_version=1.5.1
+ARG keycloak_sha1_version=0.9.0
 
-COPY ./keycloak-bcrypt-${keycloak_bcrypt_version}.jar /opt/keycloak/providers/
+EXPOSE 8080
+
+COPY ./keycloak-sha1-${keycloak_sha1_version}.jar /opt/keycloak/providers/
