+++ src/main/resources/META-INF/services/org.keycloak.credential.hash.PasswordHashProviderFactory
@@ -1,1 +1,1 @@
-com.github.leroyguillaume.keycloak.bcrypt.BCryptPasswordHashProviderFactory
+com.github.jcschaff.keycloak.sha1.SHA1PasswordHashProviderFactory
