+++ src/test/java/com/github/jcschaff/keycloak/sha1/BCryptPasswordHashProviderTest.java
@@ -1,4 +1,4 @@
-package com.github.leroyguillaume.keycloak.bcrypt;
+package com.github.jcschaff.keycloak.sha1;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -7,14 +7,14 @@ import org.keycloak.models.credential.PasswordCredentialModel;
 import static org.junit.jupiter.api.Assertions.*;
 
 class BCryptPasswordHashProviderTest {
-    private final int iterations = 10;
     private final String id = "bcrypt";
-    private final BCryptPasswordHashProvider provider = new BCryptPasswordHashProvider(id, iterations);
+    private final SHA1PasswordHashProvider provider = new SHA1PasswordHashProvider(id);
 
     @Test
     @DisplayName("Should hash the password successfully")
     void shouldHashThePasswordSuccessfully() {
         String rawPassword = "test";
+        int iterations = 1; // value is ignored
         String hashedPassword = provider.encode(rawPassword, iterations);
         PasswordCredentialModel model = PasswordCredentialModel.createFromValues(id, new byte[0], iterations, hashedPassword);
